rename testdata to exampledata
